pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-red_teaming_env"
version = "0.1.0"
description = "RedTeamEnv OpenEnv environment for adversarial LLM red teaming"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.24.0",
    "openai>=1.30.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "transformers>=4.37.0",
    "torch>=2.0.0",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "red_teaming_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["red_teaming_env", "red_teaming_env.server"]
package-dir = { "red_teaming_env" = ".", "red_teaming_env.server" = "server" }
